+++ src/components/CTRDisplay/styles.css
