+++ src/components/GraphicalInterface/GraphicalInterface.tsx
@@ -1,5 +1,18 @@
-function GraphicalInterface() {
-  return <h1>Graphical interface</h1>;
-}
+import React from 'react';
+import CRTDisplay from '@/components/CTRDisplay/CTRDisplay';
+
+const GraphicalInterface: React.FC = () => {
+  return (
+    <CRTDisplay
+      initialPowerState={true}
+      onPowerChange={isOn => console.log('Power state:', isOn)}
+    >
+      <div className="space-y-2">
+        <p>GUI</p>
+        <p className="animate-pulse">█</p>
+      </div>
+    </CRTDisplay>
+  );
+};
 
 export default GraphicalInterface;
